requirements.txt
google-adk
requests
python-dotenv
pydantic
tqdm
